Amend links so they will work.
They will NOT work on the Github repo, but should work on the Guild site where deployed. Github and Gitbook use automatic anchors guaranteed to be unique, but the Guild site uses automatic anchors that duplicate each other when 2 or more headers have the same content. The documented manual method of ID (thus anchor) definition works OFF the Github repo, but not on it. So previewing this change on Github will NOT show the document as readers will see it on the Guild site. I have deployed the document locally as a test, and there my links work as they should. They don’t jump, but rather glide, to their destinations. But the Guild repo does not contain the “_book” directory of utility files, as my standard installation does, so the behaviors may differ.